[REF] Use dedicated clientId to store the unique badge for each client
So far we used the CAmkES generated function  networkStack_rpc_enumerate_badge(0) to subtract our typical badge ID offset of 101 from all the badges to get to the index number for the client array. But as this auto-generated function is not built deterministically, we have no guarantee that the 0 index will result in the badge ID 101. We could also replace this with some fixed constant here but in the long run it seems more flexible to store the ID and derive the index through this ID when searching for the client as we can then decided freely, what badge IDs we want to store here.