[FIX] SEOS-3097: Fix clientId alignment with CAmkES client configuration
We separately assign the A) notifications and configurations  and B) RPC endpoint and dataport to the client struct. Because networkStack_rpc_enumerate_badge(i) will not guarantee the sequence in which the RPC endpoints are connected, there can be a mixup of assignments. So we need to manually assign the clientId to the notification based on the badges we assign manual via NetworkStack_PicoTcp_CLIENT_ASSIGN_BADGES.